move UI hooks to ui package
