+++ packages/common/src/index.ts
@@ -3,7 +3,3 @@ export * from "./createDevice/createDevice";
 export * from "./createEncryptionKeyFromOpaqueExportKey/createEncryptionKeyFromOpaqueExportKey";
 export * from "./createIntroductionDocumentSnapshot/createIntroductionDocumentSnapshot";
 export * from "./decryptDevice/decryptDevice";
-export * from "./useIsDesktopDevice/useIsDesktopDevice";
-export * from "./useIsEqualOrLargerThanBreakpoint/useIsEqualOrLargerThanBreakpoint";
-export * from "./useIsSmallerThanBreakpoint/useIsSmallerThanBreakpoint";
-export * from "./types";
