--- packages/common/src/types.ts
@@ -1,1 +1,0 @@
-export type Breakpoint = "xs" | "sm" | "md" | "lg" | "xl";
