+++ packages/ui/hooks/useIsEqualOrLargerThanBreakpoint/useIsEqualOrLargerThanBreakpoint.ts
@@ -1,9 +1,9 @@
 import { useWindowDimensions } from "react-native";
 import { theme } from "../../../../tailwind.config";
-import { Breakpoint } from "../types";
+import { Breakpoint } from "../../types";
 
 export const useIsEqualOrLargerThanBreakpoint = (breakpoint: Breakpoint) => {
   const { width } = useWindowDimensions();
 
-  return (width) => parseInt(theme.screens[breakpoint].replace("px"), 10);
+  return width >= parseInt(theme.screens[breakpoint].replace("px"), 10);
 };
