+++ packages/ui/hooks/useIsPermanentLeftSidebar/useIsPermanentLeftSidebar.ts
