+++ packages/ui/index.ts
@@ -31,4 +31,8 @@ export * from "./components/spinner/Spinner";
 export * from "./components/pressable/Pressable";
 export * from "./components/tooltip/Tooltip";
 
-export * from "./hooks/useIsPermanentLeftSidebar";
+export * from "./hooks/useIsDesktopDevice/useIsDesktopDevice";
+export * from "./hooks/useIsEqualOrLargerThanBreakpoint/useIsEqualOrLargerThanBreakpoint";
+export * from "./hooks/useIsPermanentLeftSidebar/useIsPermanentLeftSidebar";
+export * from "./hooks/useIsSmallerThanBreakpoint/useIsSmallerThanBreakpoint";
+export * from "./types";
