+++ packages/ui/types.ts
@@ -1,0 +1,1 @@
+export type Breakpoint = "xs" | "sm" | "md" | "lg" | "xl";
